Rename type HttpJsonFetcher -> HttpDataFetcher
This is a generic type, so there is no guarantee that the result
is JSON. The result can be any type.
