+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -7,6 +7,8 @@ export {
   mediaApi,
   isExperienceEditorActive,
   resetExperienceEditorChromes,
+  HttpDataFetcher,
+  HttpResponse,
   AxiosDataFetcher,
   AxiosDataFetcherConfig,
   LayoutService,
