+++ packages/sitecore-jss/src/axios-fetcher.ts
@@ -52,7 +52,7 @@ export class AxiosDataFetcher {
   }
 
   /**
-   * Implements a data fetcher. @see HttpJsonFetcher<T> type for implementation details/notes.
+   * Implements a data fetcher. @see HttpDataFetcher<T> type for implementation details/notes.
    * @param {string} url The URL to request; may include query string
    * @param {any} [data] Optional data to POST with the request.
    * @returns {Promise<AxiosResponse<T>>} response
