+++ packages/sitecore-jss/src/data-api.test.ts
@@ -4,12 +4,12 @@
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 import { expect } from 'chai';
-import { fetchPlaceholderData, fetchRouteData } from './dataApi';
-import { HttpJsonFetcher } from './http-request-client';
+import { fetchPlaceholderData, fetchRouteData } from './data-api';
+import { HttpDataFetcher } from './data-fetcher';
 
 // note: axios needs to use `withCredentials: true` in order for Sitecore cookies to be included in CORS requests
 // which is necessary for analytics and such
-const axiosFetcher: HttpJsonFetcher<any> = (url, data) =>
+const axiosFetcher: HttpDataFetcher<any> = (url, data) =>
   axios({
     url,
     method: data ? 'POST' : 'GET',
