+++ packages/sitecore-jss/src/dictionary-service.ts
@@ -2,7 +2,7 @@ import mcache from 'memory-cache';
 import { AxiosDataFetcher } from './axios-fetcher';
 import { fetchData } from './data-api';
 import { DictionaryPhrases, DictionaryServiceData } from './data-models';
-import { HttpJsonFetcher } from './data-fetcher';
+import { HttpDataFetcher } from './data-fetcher';
 
 export interface DictionaryService {
   /**
@@ -21,7 +21,7 @@ export type DictionaryServiceConfig = {
   getUrl(language: string): string;
   /**
    * Custom data fetcher
-   * @see HttpJsonFetcher<T>
+   * @see HttpDataFetcher<T>
    */
   dataFetcher?: HttpJsonFetcher<DictionaryServiceData>;
   /**
