+++ packages/sitecore-jss/src/layout-service.ts
@@ -2,7 +2,7 @@ import { AxiosDataFetcher, AxiosDataFetcherConfig } from './axios-fetcher';
 import { LayoutServiceData, PlaceholderData } from './data-models';
 import { fetchPlaceholderData, fetchRouteData, LayoutServiceConfig } from './data-api';
 import { GraphQLRequestClient } from './graphql-request-client';
-import { HttpJsonFetcher } from './data-fetcher';
+import { HttpDataFetcher } from './data-fetcher';
 import { AxiosRequestConfig, AxiosResponse } from 'axios';
 import { IncomingMessage, ServerResponse } from 'http';
 
@@ -26,7 +26,7 @@ export interface LayoutService {
 export type DataFetcherResolver = <T>(
   req?: IncomingMessage,
   res?: ServerResponse
-) => HttpJsonFetcher<T>;
+) => HttpDataFetcher<T>;
 
 export type RestLayoutServiceConfig = {
   /**
@@ -51,7 +51,7 @@ export type RestLayoutServiceConfig = {
   /**
    * Data fetcher resolver in order to provide custom data fetcher
    * @see DataFetcherResolver
-   * @see HttpJsonFetcher<T>
+   * @see HttpDataFetcher<T>
    * @see AxiosDataFetcher used by default
    * @param {IncomingMessage} [req] Request instance
    * @param {ServerResponse} [res] Response instance
